rule_import_web/requirements.txt
Flask==2.3.3
pandas==2.0.3
openpyxl==3.1.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0